@@ -8,9 +8,9 @@ import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../Presentation/constants/text_styles.dart';
-import '../../Presentation/utils/Text.dart';
 import '../../Presentation/utils/auth_decoration.dart';
 import '../../Presentation/utils/shimmer_effect.dart';
+import '../../lang.dart';
 
 class AuthGate extends StatefulWidget {
   const AuthGate({Key? key}) : super(key: key);
